requirements.txt
asgiref==3.4.1
astroid==2.7.3
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.7
django-rest-framework==0.1.0
djangorestframework==3.12.4
filelock==3.0.12
idna==3.2
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
platformdirs==2.3.0
pytz==2021.1
redis==3.5.3
requests==2.26.0
requests-file==1.5.1
six==1.16.0
sqlparse==0.4.1
tldextract==3.1.2
toml==0.10.2
urllib3==1.26.6
wrapt==1.12.1
